flask_app-UPLOAD-AND-EXTRACT-TEXT/flask_app-PHASE1-DB-CREATION-OK-v2.4.1/requirements.txt
poppler-utils
pdf2image
pytesseract
mysql-connector-python
flask